+++ packages/Bio/detectors.js
@@ -88,7 +88,9 @@ class BioPackageDetectors extends DG.Package {
   }
 
   /** Parts of the column name required in the column's name under the detector. It must be in lowercase. */
-  likelyColNamePartList = ['seq', 'msa', 'dna', 'rna', 'fasta', 'helm', 'sense', 'protein'];
+  likelyColNamePartList = ['seq', 'msa', 'dna', 'rna', 'fasta', 'helm', 'sense', 'protein', 'pep', 'enumerated'];
+
+  veryLikelyColNamePartList = ['peptide', 'oligo', 'sequence', 'enumerated'];
 
   peptideFastaAlphabet = new Set([
     'G', 'L', 'Y', 'S', 'E', 'Q', 'D', 'N', 'F', 'A',
@@ -152,7 +154,9 @@ class BioPackageDetectors extends DG.Package {
       const colName = col.name;
       const colNameLikely = this.likelyColNamePartList.some(
         (requiredColNamePart) => colName.toLowerCase().includes(requiredColNamePart));
-      const seqMinLength = colNameLikely ? 7 : 10;
+      const colNameVeryLikely = this.veryLikelyColNamePartList.some(
+        (requiredColNamePart) => colName.toLowerCase().includes(requiredColNamePart));
+      const seqMinLength = colNameVeryLikely ? 3 : colNameLikely ? 7 : 10;
       const maxBadRatio = colNameLikely ? 0.05 : 0.005;
 
       // Fail early
@@ -188,11 +192,11 @@ class BioPackageDetectors extends DG.Package {
         col.setTag(DG.TAGS.CELL_RENDERER, 'helm');
         return DG.SEMTYPE.MACROMOLECULE;
       }
-
+      const multiplier = colNameVeryLikely ? 1.4 : colNameLikely ? 1.2 : 1.0;
       const decoyAlphabets = [
-        ['NUMBERS', this.numbersRawAlphabet, 0.25, undefined],
-        ['SMILES', this.smilesRawAlphabet, 0.25, (seq) => seq.replaceAll()],
-        ['SMARTS', this.smartsRawAlphabet, 0.45, undefined],
+        ['NUMBERS', this.numbersRawAlphabet, 0.25 * multiplier, undefined],
+        ['SMILES', this.smilesRawAlphabet, 0.25 * multiplier, (seq) => seq.replaceAll()],
+        ['SMARTS', this.smartsRawAlphabet, 0.45 * multiplier, undefined],
       ];
 
       const candidateAlphabets = [