+++ packages/CddVaultLink/CHANGELOG.md
@@ -1,3 +1,9 @@
-# CddVaultLink changelog
+# CDD Vault Link changelog
 
-## 0.0.1 (2025-02-16)
+## 1.0.1 (2025-04-01)
+
+* Fixed error when receiving empty result
+
+## 1.0.0 (2025-04-01)
+
+Integration with [CDD Vault](https://www.collaborativedrug.com/cdd-informatics-platform) registration system.