+++ packages/Chembl/CHANGELOG.md
@@ -1,4 +1,8 @@
-# Chembl changelog
+# ChEMBL changelog
+
+## 1.4.3 (2025-04-01)
+
+* Fixed MolregnoInfo panel
 
 ## 1.4.2 (2025-03-31)