+++ packages/ClinicalCase/CHANGELOG.md
@@ -1,4 +1,8 @@
-# ClinicalCase changelog
+# Clinical Case changelog
+
+## 1.3.2 (2025-04-01)
+
+* Modified application initial view
 
 ## 1.3.1 (2025-03-31)