+++ packages/SignalsLink/CHANGELOG.md
@@ -1,3 +1,3 @@
-# SignalsLink changelog
+# Signals Link changelog
 
 ## 0.0.1 (2025-02-26)